+++ src/test/java/roomescape/admin/MissionStep4Test.java
@@ -5,7 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.ReservationRestController;
+import roomescape.controller.api.ReservationRestController;
 
 import java.lang.reflect.Field;
